code changes

+++ budgetplanner/src/main/java/com/myproject/budgetplanner/expense/ExpenseController.java
@@ -5,13 +5,19 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import com.myproject.budgetplanner.income.Income;
+import com.myproject.budgetplanner.income.IncomeNotFoundException;
 
 @RestController
 @RequestMapping("/api/expense")
@@ -30,18 +36,36 @@ public class ExpenseController {
     }
 
     @GetMapping(value = "/{id}", produces = "application/json")
-    public Optional<Expense> getExpenseById(@PathVariable long id){
-        return expenseService.getExpenseById(id);
+    public ResponseEntity<Expense> getExpenseById(@PathVariable long id){
+         Expense expense = expenseService.getExpenseById(id);
+        if (expense == null) {
+            throw new ExpenseNotFoundException("Expense not found with id: " + id);
+        }
+        return ResponseEntity.ok(expense);
+        
     }
 
     @PostMapping(produces = "application/json")
-    public Expense createExpense(@RequestBody Expense expense) throws ExpenseException{
-        return expenseService.createExpense(expense);
+    public ResponseEntity<Expense> createExpense(@RequestBody Expense expense) throws ExpenseException{
+        Expense createdExpense = expenseService.createExpense(expense);
+        return ResponseEntity.status(HttpStatus.CREATED).body(createdExpense);
+    }    
+    
+
+    @PutMapping(value = "/{id}", produces = "application/json")
+    public ResponseEntity<Expense> updateExpense(@PathVariable Long id, @RequestBody Expense updatedExpense) {
+        Expense expense = expenseService.updateExpense(id, updatedExpense);
+        return ResponseEntity.ok(expense);
     }
 
     @DeleteMapping(value = "/{id}")
-    public void deleteExpense(Long id){
+    public ResponseEntity<Void> deleteExpense(@PathVariable Long id){
+        Expense expense = expenseService.getExpenseById(id);
+        if (expense == null){
+            throw new ExpenseNotFoundException("Expense not found with id: " + id);
+        }
         expenseService.deleteExpense(id);
+        return ResponseEntity.noContent().build();
     }
 }
